+++ sandbox_town_backend/src/main/java/com/shuidun/sandbox_town_backend/service/GameMapService.java
@@ -365,7 +365,7 @@ public class GameMapService {
     }
 
     // 计算两点之间的距离
-    public double calcDistance(int x1, int y1, int x2, int y2) {
+    public double calcDistance(double x1, double y1, double x2, double y2) {
         return Math.sqrt(Math.pow(x1 - x2, 2) + Math.pow(y1 - y2, 2));
     }
 }
